Initial working code
This uses a small hard coded time range, as we need the # of entries
to be very small for initial development.

This may need parallelising, as the current serial process seems to
be inefficient and take a long time.  With the several million entries
needing to be converted, parallel may be the only feasible way.